refactor : type 위치 변경
refactor : common type 분리 및 위치 변경

refactor : store type 위치 변경 및 불필요한 파일 제거

refactor : 예약 모달 장소명 변경된 이름 적용
