+++ src/constants/calendar.ts
@@ -1,6 +1,4 @@
-import { ComponentPropsWithoutRef, PropsWithChildren } from "react";
-
-import dayjs, { type Dayjs } from "dayjs";
+import dayjs from "dayjs";
 
 export const daysOfTheWeek = ["일", "월", "화", "수", "목", "금", "토"];
 export const daysOfTheWeekByEnglish = [
@@ -30,20 +28,3 @@ export const scheduleYears = Array.from(
 );
 export const reservationYears = [dayjs().year()];
 export const months = Array.from({ length: 12 }, (_, i) => i);
-
-export interface CalendarProps {
-  startDate?: Dayjs;
-  selectedDate?: Dayjs;
-  limit?: number;
-  onChangeSelectedDate: (day: Dayjs) => void;
-}
-
-export interface CalendarItemProps
-  extends PropsWithChildren,
-    ComponentPropsWithoutRef<"td"> {
-  dayOfTheWeek: number;
-  dayData?: Dayjs;
-  type?: string;
-  isInactive?: boolean;
-  isSelected?: boolean;
-}
