+++ src/pages/Reservation/components/ReservationInfo/index.tsx
@@ -4,8 +4,8 @@ import { useFormContext, useWatch } from "react-hook-form";
 
 import Input from "@/components/Input/Input";
 import InputLabel from "@/components/Label/InputLabel";
-import type { User } from "@/dummy/user";
 import { ReservationLabel } from "@/pages/Reservation";
+import type { User } from "@/types/common";
 
 interface InfoInput extends ComponentPropsWithoutRef<"label"> {
   label: string;
