+++ src/pages/Reservation/components/ReservationTab/ReservationTabLayout/index.tsx
@@ -1,11 +1,8 @@
-import { Children, ReactNode } from "react";
+import { Children } from "react";
 
 import { ReservationLabel } from "@/pages/Reservation";
+import { PropsWithRequiredChildren } from "@/types/common";
 import { cn } from "@/utils/cn";
-
-type PropsWithRequiredChildren<P = unknown> = P & {
-  children: ReactNode;
-};
 
 interface LayoutProps extends PropsWithRequiredChildren {
   title: string;
@@ -50,16 +47,16 @@ const Bottom = ({ children, title, errorMessage }: LayoutProps) => (
       {children ? (
         <div className="flex flex-1 justify-end gap-4 text-small font-light">
           <div className="flex flex-row items-center gap-2">
-            <div className="bg-red-light h-2 w-2 rounded-full"></div>
+            <div className="h-2 w-2 rounded-full bg-red-light"></div>
             <div>예약 불가</div>
           </div>
           <div className="flex flex-row items-center gap-2">
-            <div className="bg-yellow-light h-2 w-2 rounded-full"></div>
+            <div className="h-2 w-2 rounded-full bg-yellow-light"></div>
             <div>예약 진행 중</div>
           </div>
         </div>
       ) : (
-        <div className="text-red-light text-small">{errorMessage}</div>
+        <div className="text-small text-red-light">{errorMessage}</div>
       )}
     </div>
     {children}
