--- src/store/store.js
@@ -1,9 +1,0 @@
-import { create } from "zustand";
-
-const useStore = create(set => ({
-  bears: 0,
-  increasePopulation: () => set(state => ({ bears: state.bears + 1 })),
-  removeAllBears: () => set({ bears: 0 }),
-}));
-
-export default useStore;
